fix software reset by writing in SYSRESETQ register when a reset is requested (#73)
Some debugger software (e.g. Keil) does not perform a software reset when
resetting the target. When this option enabled, a soft reset is attempted
when DAP_ResetTarget() is performed. This is done by writing to the
SYSRESETREQ field of the AIRCR register in the Cortex-M architecture.

This should work for ARMv6-m, ARMv7-m and ARMv8-m architecture. However,
there is no guarantee that the reset operation will be executed correctly.

Only available for SWD.

Links: https://github.com/windowsair/wireless-esp8266-dap/pull/73
---------

Co-authored-by: windowsair <msdn_02@sina.com>